Relocate file
